lifelens/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
google-generativeai==0.8.5
qdrant-client==1.15.1
httpx==0.28.1
